% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocboost_assemble.R
\name{colocboost_assemble}
\alias{colocboost_assemble}
\title{Main function for colocboost post aggregating analysis}
\usage{
colocboost_assemble(
  cb_obj,
  coverage = 0.95,
  weight_fudge_factor = 1.5,
  check_null = 0.1,
  check_null_method = "profile",
  check_null_max = 2e-05,
  dedup = TRUE,
  overlap = TRUE,
  n_purity = 100,
  min_abs_corr = 0.5,
  sec_coverage_thresh = 0.8,
  median_abs_corr = NULL,
  min_cluster_corr = 0.8,
  median_cos_abs_corr = 0.8,
  weaker_effect = TRUE,
  merge_cos = TRUE,
  tol = 1e-09,
  output_level = 1
)
}
\description{
Main function for colocboost post aggregating analysis
}
\details{
The following functions are included in the post-hoc analysis:

Colocalization signal - \code{colocboost_assemble_cos} - identify the colocalized confidence sets and the corresponding causal configurations.

Un-colocalization signal - \code{colocboost_assemble_ucos} - identify the causal confidence sets for each outcome only.

Add-hoc merge_cos functions including

\item{merge_coloc_single}{merge the colocalized sets and the single causal set if pass the \code{median_cos_abs_corr}}
\item{merge_single}{merge the single causal sets for different outcomes if pass the \code{median_cos_abs_corr}}

Refine of the colocalization sets (TO-DO-LIST)

Summary of the colocboost results and get the output of colocboost (TO-DO-LIST)
}
