% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocboost_check_update_jk.R
\name{colocboost_check_update_jk}
\alias{colocboost_check_update_jk}
\title{Model selection scheme in ColocBoost}
\usage{
colocboost_check_update_jk(
  cb_model,
  cb_model_para,
  cb_data,
  prioritize_jkstar = TRUE,
  jk_equiv_corr = 0.8,
  jk_equiv_loglik = 1,
  func_compare = "min_max",
  coloc_thresh = 0.1
)
}
\value{
update_status and real_update_jk for each trait
}
\description{
Model selection scheme in ColocBoost
}
\details{
Model selection scheme in proximity smoothed gradient boosting algorithm to select update traits and update variants at each iteration.
}
