% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocboost_init.R
\name{colocboost_inits}
\alias{colocboost_inits}
\title{Set of Internal functions for initial colocboost objects}
\usage{
colocboost_init_data(X, Y, dict_YX, Z, LD, N_sumstat, dict_sumstatLD, Var_y, SeBhat)
colocboost_init_model(cb_data)
colocboost_init_para(cb_data, cb_model)
}
\description{
The \code{colocboost_inits} function provides an interface for initializing different colocboost objects, including data, model,
and model parameters. This documentation serves as a summary for all related initialization functions.
}
\details{
The following functions are included in this set:
\code{colocboost_init_data} initial colocboost data object.
\code{colocboost_init_model} initial colocboost model object.
\code{colocboost_init_para} initial colocboost model global parameters object.

These functions are not exported individually and are accessed via \code{colocboost_inits}.
}
\keyword{cb_objects}
