% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocboost_inference.R
\name{colocboost_post_inference}
\alias{colocboost_post_inference}
\title{Set of functions for post inferences from ColocBoost}
\usage{
colocboost_post_inference()
}
\description{
The \code{colocboost_post_inference} functions access basic properties inferences from a fitted ColocBoost model. This documentation serves as a summary for all related post-inference functions.
}
\details{
The following functions are included in this set:
\code{get_abc} get the colocalization summary table with or without the specific outcomes.

These functions are not exported individually and are accessed via \code{colocboost_post_inference}.
}
\keyword{cb_post_inference}
