% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocboost_posthoc.R
\name{colocboost_posthoc}
\alias{colocboost_posthoc}
\title{Main function for colocboost post-hoc analysis}
\usage{
colocboost_posthoc(
  cb_obj,
  coverage = 0.95,
  func_intw = "fun_R",
  alpha = 1.5,
  check_null = 0.1,
  check_null_method = "profile",
  check_null_max = 2e-05,
  dedup = TRUE,
  overlap = TRUE,
  n_purity = 100,
  min_abs_corr = 0.5,
  coverage_singlew = 0.8,
  median_abs_corr = NULL,
  between_cluster = 0.8,
  between_purity = 0.8,
  weaker_ucos = TRUE,
  merging = TRUE,
  tol = 1e-09,
  output_level = 1
)
}
\description{
Main function for colocboost post-hoc analysis
}
\details{
The following functions are included in the post-hoc analysis:

Colocalization signal - \code{colocboost_posthoc_coloc} - identify the colocalized confidence sets and the corresponding causal configurations.

Un-colocalization signal - \code{colocboost_posthoc_noncoloc} - identify the causal confidence sets for each outcome only.

Add-hoc merging functions including

\itemize{
\item{merge_coloc_single}{merge the colocalized sets and the single causal set if pass the \code{between_purity}}
\item{merge_single}{merge the single causal sets for different outcomes if pass the \code{between_purity}}
}

Refine of the colocalization sets (TO-DO-LIST)

Summary of the colocboost results and get the output of colocboost (TO-DO-LIST)
}
