% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocboost_update.R
\name{colocboost_update}
\alias{colocboost_update}
\title{Joint boosting algorithm in ColocBoost}
\usage{
colocboost_update(
  cb_model,
  cb_model_para,
  cb_data,
  tau = 0.01,
  learning_rate_decay = 1,
  func_simplex = "z2z",
  lambda = 0.5,
  lambda_target_outcome = 1,
  LD_free = FALSE,
  dynamic_learning_rate = TRUE
)
}
\value{
colocboost object after gradient boosting update
}
\description{
Joint boosting algorithm in ColocBoost
}
\details{
The gradient boosting algorithm for multiple outcomes
}
