% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocboost_workhorse.R
\name{colocboost_workhorse}
\alias{colocboost_workhorse}
\title{Workhorse of colocboost}
\usage{
colocboost_workhorse(
  cb_data,
  M = NULL,
  step = 0.01,
  tau = 0.01,
  prioritize_jkstar = TRUE,
  decayrate = 1,
  func_prior = "z2z",
  lambda = 0.5,
  lambda_target = 1,
  jk_equiv_cor = 0.8,
  jk_equiv_loglik = 1,
  stop_thresh = 1e-06,
  func_multicorrection = "lfdr",
  stop_null = 0.05,
  multicorrection_max = 1,
  multicorrection_cut = 1,
  ash_prior = "normal",
  p.adjust.methods = "fdr",
  func_compare = "min_max",
  coloc_thres = 0.1,
  LD_obj = FALSE,
  target_idx = NULL,
  traits_names = NULL
)
}
\description{
Workhorse of colocboost
}
\details{
The following functions are included in the gradient boosting iterations:

Step 1: \code{boost_check_update_jk} selection scheme to choose traits need to be updated in each iteration.

Step 2: \code{boost_joint} core function of the proximity smoothed gradient boosting update.

Step 3: check the stop criteria for each trait.
}
