% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocboost_utils.R
\name{get_merge_ordered_with_indices}
\alias{get_merge_ordered_with_indices}
\title{Pure R implementation (fallback)}
\usage{
get_merge_ordered_with_indices(vector_list)
}
\description{
Pure R implementation (fallback)
}
\keyword{internal}
