% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Ind_5traits}
\alias{Ind_5traits}
\title{Individual level data for 5 traits}
\format{
\subsection{\code{Ind_5traits}}{

A list with 3 elements
\describe{
\item{X}{List of genotype matrices}
\item{Y}{List of traits}
\item{true_effect_variants}{List of causal variants}
}
}
}
\source{
The Ind_5traits dataset contains 5 simulated phenotypes alongside corresponding genotype matrices.
The dataset is specifically designed for evaluating and demonstrating the capabilities of ColocBoost in multi-trait colocalization analysis
with individual-level data. See Cao etc. 2025 for details.
Due to the file size limitation of CRAN release, this is a subset of simulated data.
See full dataset in colocboost paper repo \url{https://github.com/StatFunGen/colocboost-paper}.
}
\usage{
Ind_5traits
}
\description{
An example dataset with simulated genotypes and traits for 5 traits
}
\seealso{
Other colocboost_data:
\code{\link{Ambiguous_Colocalization}},
\code{\link{Heterogeneous_Effect}},
\code{\link{Non_Causal_Strongest_Marginal}},
\code{\link{Sumstat_5traits}},
\code{\link{Weaker_GWAS_Effect}}
}
\concept{colocboost_data}
\keyword{datasets}
