% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sumstat_5traits}
\alias{Sumstat_5traits}
\title{Summary level data for 5 traits}
\format{
\subsection{\code{Sumstat_5traits}}{

A list with 2 elements
\describe{
\item{sumstat}{Summary statistics for 5 traits}
\item{true_effect_variants}{List of causal variants}
}
}
}
\source{
The Sumstat_5traits dataset contains 5 simulated summary statistics,
where it is directly derived from the Ind_5traits dataset using marginal association.
The dataset is specifically designed for evaluating and demonstrating the capabilities of ColocBoost
in multi-trait colocalization analysis with summary association data. See Cao etc. 2025 for details.
Due to the file size limitation of CRAN release, this is a subset of simulated data.
See full dataset in colocboost paper repo \url{https://github.com/StatFunGen/colocboost-paper}.
}
\usage{
Sumstat_5traits
}
\description{
An example dataset with simulated statistics for 5 traits
}
\seealso{
Other colocboost_data:
\code{\link{Ambiguous_Colocalization}},
\code{\link{Heterogeneous_Effect}},
\code{\link{Ind_5traits}},
\code{\link{Non_Causal_Strongest_Marginal}},
\code{\link{Weaker_GWAS_Effect}}
}
\concept{colocboost_data}
\keyword{datasets}
