% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Weaker_GWAS_Effect}
\alias{Weaker_GWAS_Effect}
\title{Individual level data for 2 traits and 2 causal variants with weaker effects for focal trait}
\format{
\subsection{\code{Weaker_GWAS_Effect}}{

A list with 3 elements
\describe{
\item{X}{List of genotype matrices}
\item{Y}{List of traits}
\item{variant}{indices of two causal variants}
}
}
}
\source{
The Weaker_GWAS_Effect dataset contains 2 simulated phenotypes alongside corresponding genotype matrices.
There are two causal variants, one of which has a weaker effect on the focal trait compared to the other trait.
Due to the file size limitation of CRAN release, this is a subset of simulated data to generate Figure 2b in Cao etc. 2025.
See full dataset in colocboost paper repo \url{https://github.com/StatFunGen/colocboost-paper}.
}
\usage{
Weaker_GWAS_Effect
}
\description{
An example dataset with simulated genotypes and traits for 2 traits and 2 common causal variants with heterogeneous effects
}
\seealso{
Other colocboost_data:
\code{\link{Ambiguous_Colocalization}},
\code{\link{Heterogeneous_Effect}},
\code{\link{Ind_5traits}},
\code{\link{Non_Causal_Strongest_Marginal}},
\code{\link{Sumstat_5traits}}
}
\concept{colocboost_data}
\keyword{datasets}
