% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Ambiguous_Colocalization}
\alias{Ambiguous_Colocalization}
\title{A real data example includes an ambiguous colocalization between eQTL and GWAS}
\format{
\subsection{\code{Ambiguous_Colocalization}}{

A list with 2 elements
\describe{
\item{ColocBoost_Results}{A \code{colocboost} output object}
\item{SuSiE_Results}{Two \code{susie} output object for eQTL and GWAS}
\item{COLOC_V5_Results}{A \code{coloc} output object}
}
}
}
\source{
The Ambiguous_Colocalization dataset contains a real data example from one of our real data applications,
which shows an ambiguous colocalization between eQTL and GWAS.
The dataset is specifically designed for evaluating and demonstrating the capabilities of ColocBoost in real data applications.
See details in tutorial vignette \url{https://statfungen.github.io/colocboost/articles/index.html}.
}
\usage{
Ambiguous_Colocalization
}
\description{
An example result from one of our real data applications, which shows an ambiguous colocalization between eQTL and GWAS.
}
\seealso{
Other colocboost_data: 
\code{\link{Heterogeneous_Effect}},
\code{\link{Ind_5traits}},
\code{\link{Non_Causal_Strongest_Marginal}},
\code{\link{Sumstat_5traits}},
\code{\link{Weaker_GWAS_Effect}}
}
\concept{colocboost_data}
\keyword{datasets}
