% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocboost.R
\name{colocboost}
\alias{colocboost}
\title{ColocBoost: A gradient boosting informed multi-omics colocalization method}
\usage{
colocboost(
  X = NULL,
  Y = NULL,
  sumstat = NULL,
  LD = NULL,
  dict_YX = NULL,
  dict_sumstatLD = NULL,
  traits_names = NULL,
  effect_est = NULL,
  effect_se = NULL,
  effect_n = NULL,
  M = NULL,
  stop_thresh = 1e-06,
  step = 0.01,
  decayrate = 1,
  tau = 0.01,
  prioritize_jkstar = TRUE,
  func_compare = "min_max",
  jk_equiv_cor = 0.8,
  jk_equiv_loglik = 1,
  coloc_thres = 0.1,
  intercept = TRUE,
  standardize = TRUE,
  coverage = 0.95,
  between_cluster = 0.8,
  dedup = TRUE,
  overlap = TRUE,
  n_purity = 100,
  min_abs_corr = 0.5,
  median_abs_corr = NULL,
  between_purity = 0.8,
  tol = 1e-09,
  merging = TRUE,
  coverage_singlew = 0.8,
  lambda = 0.5,
  lambda_target = 1,
  func_intw = "fun_R",
  alpha = 1.5,
  func_prior = "LD_z2z",
  func_multicorrection = "lfdr",
  target_idx = NULL,
  target_variants = TRUE,
  overlap_varaints = FALSE,
  stop_null = 1,
  multicorrection_max = 1,
  multicorrection_cut = 1,
  ash_prior = "normal",
  p.adjust.methods = "fdr",
  check_null = 0.1,
  check_null_method = "profile",
  check_null_max = 0.02,
  residual_correlation = NULL,
  LD_obj = FALSE,
  weaker_ucos = TRUE,
  output_level = 1
)
}
\arguments{
\item{X}{A list of genotype matrices for different traits, or a single matrix if all traits share the same genotypes.
Each matrix should have column names, if sample sizes and variants possibly differing across matrices.}

\item{Y}{A list of vectors of traits or an N by L matrix if it is considered for the same X and multiple traits.}

\item{sumstat}{A list of data.frames of summary statistics.
The coloumns of data.frame should include either \code{z} or \code{beta}/\code{sebeta}.
\code{n} is the sample size for the summary statistics, it is highly recommendation to provide.
\code{variants} is required if sumstat for different traits do not have the same number of variants.
\code{var_y} is the variance of phenotype (default is 1 meaning that the Y is in the \dQuote{standarized} scale).}

\item{LD}{A list of correlation matrix indicating the LD matrix for each genotype. It also could be a single matrix if all sumstats were
obtained from the same gentoypes.}

\item{dict_YX}{A L by 2 matrix of dictionary for \code{X} and \code{Y} if there exist subsets of traits corresponding to the same X matrix.
The first column should be 1:L for L traits. The second column should be the index of \code{X} corresponding to the trait.
The innovation: do not provide the same matrix in \code{X} to reduce the computational burden.}

\item{dict_sumstatLD}{A L by 2 matrix of dictionary for \code{sumstat} and \code{LD} if there exist subsets of traits corresponding to the same sumstat.
The first column should be 1:L for L sumstat The second column should be the index of \code{LD} corresponding to the sumstat.
The innovation: do not provide the same matrix in \code{LD} to reduce the computational burden.}

\item{traits_names}{The names of traits, which has the same order for Y.}

\item{M}{The maximum number of gradient boosting iterations. If the number of traits are large, it will be automatically increased to a larger number.}

\item{stop_thresh}{The stop criterion for overall profile loglikelihood function.}

\item{step}{The minimum step size (learning rate) for updating in each iteration.}

\item{decayrate}{The decayrate for step size. If the objective function is large at the early iterations,
we need to have the higher step size to improve the computational efficiency.}

\item{tau}{The smooth parameter for proximity smoothed weights for the best update jk-star.}

\item{prioritize_jkstar}{When \code{prioritize_jkstar = TRUE}, the selected traits will prioritize best update j_k^star in boosting iteration.}

\item{jk_equiv_cor}{The LD cutoff between overall best update jk-star and marginal best update jk-l for lth trait}

\item{jk_equiv_loglik}{The change of loglikelihood cutoff between overall best update jk-star and marginal best update jk-l for lth trait}

\item{coloc_thres}{The cutoff of checking if the best update jk-star is the potential causal variant for trait l if jk-l is not similar to jk-star.}

\item{intercept}{If \code{intercept = TRUE}, the intercept is fitted. Setting \code{intercept = FALSE} is generally not recommended.}

\item{standardize}{If \code{standardize = TRUE}, standardize the columns of genotype and traits to unit variance.}

\item{coverage}{A number between 0 and 1 specifying the \dQuote{coverage} of the estimated confidence sets (default is 0.95).}

\item{between_cluster}{The small correlation for the weights distributions across different iterations to be decided having only one cluster.}

\item{dedup}{If \code{dedup = TRUE}, the duplicate confidence sets will be removed in the post-processing.}

\item{overlap}{If \code{overlap = TRUE}, the overlapped confidence sets will be removed in the post-processing.}

\item{n_purity}{The maximum number of confidence set (CS) variables used in calculating the correlation (\dQuote{purity}) statistics.
When the number of variables included in the CS is greater than this number, the CS variables are randomly subsampled.}

\item{min_abs_corr}{Minimum absolute correlation allowed in a confidence set. The default is 0.5 corresponding to a squared correlation of 0.25,
which is a commonly used threshold for genotype data in genetic studies.}

\item{median_abs_corr}{An alternative "purity" threshold for the CS. Median correlation between pairs of variables in a CS less than this
threshold will be filtered out and not reported. When both min_abs_corr and median_abs_corr are set,
a CS will only be removed if it fails both filters. Default set to NULL but it is recommended to set it to 0.8 in practice.}

\item{between_purity}{Minimum absolute correlation allowed to merge multiple colocalized sets. The default is 0.8 corresponding to a stringent threshold
to merge colocalized sets, which may resulting in a huge set.}

\item{tol}{A small, non-negative number specifying the convergence tolerance for checking the overlap of the variants in different sets.}

\item{merging}{When \code{merging = TRUE}, the sets for only one trait will be merged if passed the \code{between_purity}.}

\item{func_multicorrection}{The alternative method to check the stop criteria. When \code{func_multicorrection = "lfdr"}, boosting iterations will be stopped
if the local FDR for all variants are greater than \code{lfsr_max}.}

\item{target_idx}{The index of the target trait}

\item{stop_null}{The cutoff of nominal p-value when \code{func_multicorrection = "Z"}.}

\item{ash_prior}{The prior distribution for calculating lfsr when \code{func_multicorrection = "lfsr"}.}

\item{check_null}{The cut off value for change conditional objective function. Default is 0.1.}

\item{check_null_method}{The metric to check the null sets. Default is "profile"}

\item{check_null_max}{The smallest value of change of profile loglikelihood for each trait.}

\item{residual_correlation}{The residual correlation based on the sample overlap, it is diagonal if it is NULL.}

\item{LD_obj}{When \code{LD_obj = FALSE}, objective fuunction doesn't include LD information.}

\item{output_level}{When \code{output_level = 2}, return the entire Colocboost model to diagnostic results (more space).}

\item{compare_method}{The criterion when we update jk-star (default is "min_max").}

\item{w_cset}{The integrated weight method. The default is "fun_R", indicating the same log-scale for different colocalized traits.}

\item{prob_style}{The penalty \eqn{\delta} for smoothing the weights. Default is "z2z" is the elastic net for z-score and also weighted by LD.}

\item{target_prop}{The proportion of the importance of the target trait. It is not a recommendation that setting up a value greater than 0.5.}

\item{lfsr_max}{The cutoff of the smallest FDR for pre-filtering the traits when \code{func_multicorrection = "lfdr"} or \code{func_multicorrection = "lfsr"}.}

\item{lfsr_cut}{The cutoff of the smallest FDR for stop criteria when \code{func_multicorrection = "lfdr"} or \code{func_multicorrection = "lfsr"}.}
}
\value{
A \code{"colocboost"} object with some or all of the following elements:

\item{coloc_results}{A object with all information for colocalization results.}

\item{vcp}{The variant colocalized probability for each variant.}
}
\description{
\code{colocboost} implements a proximity-smoothed gradient boosting approach for multi-trait colocalization at gene loci,
accommodating multiple causal variants. This method, introduced by Cao et al. (2024), is particularly suited for scaling
to large datasets involving numerous molecular quantitative traits and disease traits.
In brief, this function fits a multiple linear regression model \eqn{Y = XB + E} in matrix form.
}
\details{
The function \code{colocboost} implements the proximity smoothed gradient boosting method from Cao et al (2024).
The option \code{refine = TRUE} implements an additional step to help merge the confidence sets with small \code{between_putiry}
(default is 0.8) but within the same locus. This step addresses potential instabilities in linkage disequilibrium (LD) estimation
that may arise from small sample sizes or discrepancies in minor allele frequencies (MAF) across different confidence sets.
}
\examples{
colocboost(X=X, Y=Y)

}
