% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocboost_output.R
\name{colocboost_get_methods}
\alias{colocboost_get_methods}
\title{Set of functions for post inferences from ColocBoost}
\usage{
colocboost_get_methods()
}
\description{
The \code{colocboost_get_methods} functions access basic properties inferences from a fitted ColocBoost model. This documentation serves as a summary for all related post-inference functions.
}
\details{
The following functions are included in this set:
\code{get_summary_table} get the colocalization summary table with or without the specific outcomes.

These functions are not exported individually and are accessed via \code{colocboost_get_methods}.
}
\keyword{cb_post_inference}
