% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocboost_inference.R
\name{get_cormat}
\alias{get_cormat}
\title{A fast function to calculate correlation matrix (LD matrix) from individual level data}
\usage{
get_cormat(X, intercepte = FALSE)
}
\arguments{
\item{X}{A matrix of individual level data.}

\item{intercepte}{A logical value indicating whether to include an intercept in the model. Default is FALSE.}
}
\value{
A correlation matrix (LD matrix).
}
\description{
This function calculates the correlation matrix (LD matrix) from individual level data.
}
\keyword{cb_post_inference}
