% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocboost_update.R
\name{colocboost_update}
\alias{colocboost_update}
\title{Joint boosting algorithm in ColocBoost}
\usage{
colocboost_update(
  cb_model,
  cb_model_para,
  cb_data,
  tau = 0.01,
  decayrate = 1,
  func_prior = "z2z",
  lambda = 0.5,
  lambda_target = 1,
  LD_obj = FALSE
)
}
\value{
colocboost object after gradient boosting update
}
\description{
Joint boosting algorithm in ColocBoost
}
\details{
The gradient boosting algorithm for multiple traits
}
