% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocboost_workhorse.R
\name{colocboost_workhorse}
\alias{colocboost_workhorse}
\title{Workhorse of colocboost}
\usage{
colocboost_workhorse(
  cb_data,
  M = NULL,
  tau = 0.01,
  prioritize_jkstar = TRUE,
  learning_rate_init = 0.01,
  learning_rate_decay = 1,
  func_simplex = "z2z",
  lambda = 0.5,
  lambda_target_outcome = 1,
  jk_equiv_corr = 0.8,
  jk_equiv_loglik = 1,
  stop_thresh = 1e-06,
  func_multi_test = "lfdr",
  stop_null = 0.05,
  multi_test_max = 1,
  multi_test_thresh = 1,
  ash_prior = "normal",
  p.adjust.methods = "fdr",
  func_compare = "min_max",
  coloc_thresh = 0.1,
  LD_free = FALSE,
  dynamic_learning_rate = TRUE,
  target_outcome_idx = NULL,
  outcome_names = NULL
)
}
\description{
Workhorse of colocboost
}
\details{
The following functions are included in the gradient boosting iterations:

Step 1: \code{colocboost_check_update_jk} selection scheme to choose outcomes need to be updated in each iteration.

Step 2: \code{colocboost_update} core function of the proximity smoothed gradient boosting update.

Step 3: check the stop criteria for each outcome.

There is a version for LD free version with one causal assumption, implemented in \code{colocboost_one_causal}.
}
