% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocboost_output.R
\name{get_cos}
\alias{get_cos}
\title{Extract CoS simply change the coverage without checking purity}
\usage{
get_cos(cb_output, coverage = 0.95)
}
\arguments{
\item{cb_output}{Output object from \code{colocboost} analysis}

\item{coverage}{A number between 0 and 1 specifying the \dQuote{coverage} of the estimated colocalization confidence sets (CoS) (default is 0.95).}
}
\value{
A list of indices of variables in each CoS.
}
\description{
\code{get_cos} get the colocalization confidence sets (CoS) with different coverage.
}
\seealso{
Other colocboost_utilities: 
\code{\link{get_cormat}()}
}
\concept{colocboost_utilities}
