% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocboost_output.R
\name{get_cos_summary}
\alias{get_cos_summary}
\title{Get colocalization summary table from a ColocBoost output.}
\usage{
get_cos_summary(
  cb_output,
  outcome_names = NULL,
  interest_outcome = NULL,
  gene_name = NULL
)
}
\arguments{
\item{cb_output}{Output object from \code{colocboost} analysis}

\item{outcome_names}{Optional vector of names of outcomes, which has the same order as Y in the original analysis.}

\item{interest_outcome}{Optional vector specifying a subset of outcomes from \code{outcome_names} to focus on. When provided, only colocalization events that include at least one of these outcomes will be returned.}

\item{gene_name}{Optional character string. When provided, adds a column with this gene name to the output table for easier filtering in downstream analyses.}
}
\value{
A summary table for colocalization events with the following columns:
\item{target_outcome}{The target outcome being analyzed if exists. Otherwise, it is \code{FALSE}.}
\item{colocalized_outcomes}{Colocalized outcomes for colocalization confidence set (CoS) }
\item{cos_id}{Unique identifier for colocalization confidence set (CoS) }
\item{purity}{Minimum absolute correlation of variables with in colocalization confidence set (CoS) }
\item{top_variable}{The variable with highest variant colocalization probability (VCP) }
\item{top_variable_vcp}{Variant colocalization probability for the top variable}
\item{cos_npc}{Normalized probability of colocalization}
\item{min_npc_outcome}{Minimum normalized probability of colocalized traits}
\item{n_variables}{Number of variables in colocalization confidence set (CoS)}
\item{colocalized_index}{Indices of colocalized variables}
\item{colocalized_variables}{List of colocalized variables}
\item{colocalized_variables_vcp}{Variant colocalization probabilities for all colocalized variables}
}
\description{
\code{get_cos_summary} get the colocalization summary table with or without the outcomes of interest.
}
\examples{
get_cos_summary(cb_output)
get_cos_summary(cb_output, interest_outcome = c("Y1", "Y2"))

}
\keyword{cb_get_functions}
