% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocboost_output.R
\name{get_ucos_summary}
\alias{get_ucos_summary}
\title{Get trait-specific summary table from a ColocBoost output.}
\source{
See detailed instructions in our tutorial portal:
\url{https://statfungen.github.io/colocboost/articles/Interpret_ColocBoost_Output.html}
}
\usage{
get_ucos_summary(cb_output, outcome_names = NULL, region_name = NULL)
}
\arguments{
\item{cb_output}{Output object from \code{colocboost} analysis}

\item{outcome_names}{Optional vector of names of outcomes, which has the same order as Y in the original analysis.}

\item{region_name}{Optional character string. When provided, adds a column with this gene name to the output table for easier filtering in downstream analyses.}
}
\value{
A summary table for trait-specific, uncolocalized associations with the following columns:
\item{outcomes}{Outcome being analyzed}
\item{ucos_id}{Unique identifier for trait-specific confidence sets}
\item{purity}{Minimum absolute correlation of variables within trait-specific confidence sets}
\item{top_variable}{The variable with highest variant-level probability of association (VPA)}
\item{top_variable_vpa}{Variant-level probability of association (VPA) for the top variable}
\item{ucos_npc}{Normalized probability of causal association for the trait-specific confidence set}
\item{n_variables}{Number of variables in trait-specific confidence set}
\item{ucos_index}{Indices of variables in the trait-specific confidence set}
\item{ucos_variables}{List of variables in the trait-specific confidence set}
\item{ucos_variables_vpa}{Variant-level probability of association (VPA) for all variables in the confidence set}
\item{region_name}{Region name if provided through the region_name parameter}
}
\description{
\code{get_ucos_summary} produces a trait-specific summary table for uncolocalized (single-trait)
associations from ColocBoost results. This is particularly useful for examining trait-specific signals
or for summarizing results from single-trait FineBoost analyses.
}
\examples{
# colocboost example with single trait analysis
set.seed(1)
N <- 1000
P <- 100
# Generate X with LD structure
sigma <- 0.9^abs(outer(1:P, 1:P, "-"))
X <- MASS::mvrnorm(N, rep(0, P), sigma)
colnames(X) <- paste0("SNP", 1:P)
L <- 1  # Only one trait for single-trait analysis
true_beta <- matrix(0, P, L)
true_beta[10, 1] <- 0.5 # SNP10 affects the trait
true_beta[80, 1] <- 0.2 # SNP11 also affects the trait but with lower effect
Y <- X \%*\% true_beta + rnorm(N, 0, 1)
res <- colocboost(X = X, Y = Y, output_level = 2)
# Get the trait-specifc effect summary
get_ucos_summary(res)

}
\seealso{
Other colocboost_inference: 
\code{\link{get_cos_summary}()},
\code{\link{get_robust_colocalization}()}
}
\concept{colocboost_inference}
