% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Heterogeneous_Effect}
\alias{Heterogeneous_Effect}
\title{Individual level data for 2 traits and 2 causal variants with heterogeneous effects}
\format{
\subsection{\code{Heterogeneous_Effect}}{

A list with 3 elements
\describe{
\item{X}{List of genotype matrices}
\item{Y}{List of traits}
\item{variant}{Incides of two causal variants}
}
}
}
\source{
The Heterogeneous_Effect dataset contains 2 simulated phenotypes alongside corresponding genotype matrices.
There are two causal variants, both of which have heterogeneous effects on two traits.
See Figure 2b in Cao et. al. 2025 for details.
}
\usage{
Heterogeneous_Effect
}
\description{
An example dataset with simulated genotypes and traits for 2 traits and 2 common causal variants with heterogeneous effects
}
\seealso{
Other colocboost_data: 
\code{\link{Ambiguous_Colocalization}},
\code{\link{Ind_5traits}},
\code{\link{Non_Causal_Strongest_Marginal}},
\code{\link{Sumstat_5traits}},
\code{\link{Weaker_GWAS_Effect}}
}
\concept{colocboost_data}
\keyword{datasets}
