% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Ind_5traits}
\alias{Ind_5traits}
\title{Individual level data for 5 traits}
\format{
\subsection{\code{Ind_5traits}}{

A list with 3 elements
\describe{
\item{X}{List of genotype matrices}
\item{Y}{List of outcomes}
\item{TrueCausalVariants}{List of causal variants}
}
}
}
\source{
See Cao et. al. 2025 for details. TO-DO-LIST
}
\usage{
Ind_5traits
}
\description{
An example dataset with simulated genotypes and outcomes for 5 traits
}
\seealso{
Other colocboost_data: 
\code{\link{Heterogeneous_Effect}},
\code{\link{Non_Causal_Strongest_Marginal}},
\code{\link{Sumstat_5traits}},
\code{\link{Weaker_GWAS_Effect}}
}
\concept{colocboost_data}
\keyword{datasets}
