% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Ind_5traits}
\alias{Ind_5traits}
\title{Individual level data for 5 traits}
\format{
\subsection{\code{Ind_5traits}}{

A list with 3 elements
\describe{
\item{X}{List of genotype matrices}
\item{Y}{List of outcomes}
\item{TrueCausalVariants}{List of causal variants}
}
}
}
\source{
FINISH
}
\usage{
Ind_5traits
}
\description{
An example dataset with simulated genotypes and outcomes for 5 traits
}
\keyword{datasets}
