% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Non_Causal_Strongest_Marginal}
\alias{Non_Causal_Strongest_Marginal}
\title{Summary level data for 5 traits}
\format{
\subsection{\code{Non_Causal_Strongest_Marginal}}{

A list with 3 elements
\describe{
\item{X}{List of genotype matrices}
\item{Y}{List of outcomes}
\item{TrueCausalVariants}{List of causal variants}
}
}
}
\source{
See Cao et. al. 2025 for details. TO-DO-LIST
}
\usage{
Non_Causal_Strongest_Marginal
}
\description{
An example dataset with simulated statistics and LD for 5 traits
}
\keyword{datasets}
