% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Non_Causal_Strongest_Marginal}
\alias{Non_Causal_Strongest_Marginal}
\title{Individual level data for 2 traits and 2 causal variants, but the strongest marginal association is not causal}
\format{
\subsection{\code{Non_Causal_Strongest_Marginal}}{

A list with 3 elements
\describe{
\item{X}{List of genotype matrices}
\item{Y}{List of traits}
\item{variant}{indices of two causal variants}
}
}
}
\source{
The Non_Causal_Strongest_Marginal dataset contains 2 simulated phenotypes alongside corresponding genotype matrices.
There are two causal variants, but the strongest marginal association is not a causal variant.
Due to the file size limitation of CRAN release, this is a subset of simulated data to generate Figure 2b in Cao etc. 2025.
See full dataset in colocboost paper repo \url{https://github.com/StatFunGen/colocboost-paper}.
}
\usage{
Non_Causal_Strongest_Marginal
}
\description{
An example dataset with simulated genotypes and traits for 2 traits and 2 common causal variants, but the strongest marginal association is not causal variant.
}
\seealso{
Other colocboost_data: 
\code{\link{Ambiguous_Colocalization}},
\code{\link{Heterogeneous_Effect}},
\code{\link{Ind_5traits}},
\code{\link{Sumstat_5traits}},
\code{\link{Weaker_GWAS_Effect}}
}
\concept{colocboost_data}
\keyword{datasets}
