% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sumstat_5traits}
\alias{Sumstat_5traits}
\title{Summary level data for 5 traits}
\format{
\subsection{\code{Sumstat_5traits}}{

A list with 3 elements
\describe{
\item{sumstat}{Summary statistics for 5 traits}
\item{LD}{Matrix of LD between variants}
\item{TrueCausalVariants}{List of causal variants}
}
}
}
\source{
FINISH
}
\usage{
Sumstat_5traits
}
\description{
An example dataset with simulated statistics and LD for 5 traits
}
\keyword{datasets}
