% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sumstat_5traits}
\alias{Sumstat_5traits}
\title{Summary level data for 5 traits}
\format{
\subsection{\code{Sumstat_5traits}}{

A list with 2 elements
\describe{
\item{sumstat}{Summary statistics for 5 traits}
\item{TrueCausalVariants}{List of causal variants}
}
}
}
\source{
See Cao et. al. 2025 for details. TO-DO-LIST
}
\usage{
Sumstat_5traits
}
\description{
An example dataset with simulated statistics and LD for 5 traits
}
\seealso{
Other colocboost_data: 
\code{\link{Heterogeneous_Effect}},
\code{\link{Ind_5traits}},
\code{\link{Non_Causal_Strongest_Marginal}},
\code{\link{Weaker_GWAS_Effect}}
}
\concept{colocboost_data}
\keyword{datasets}
