% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Weaker_GWAS_Effect}
\alias{Weaker_GWAS_Effect}
\title{Summary level data for 5 traits}
\format{
\subsection{\code{Weaker_GWAS_Effect}}{

A list with 3 elements
\describe{
\item{X}{List of genotype matrices}
\item{Y}{List of outcomes}
\item{TrueCausalVariants}{List of causal variants}
}
}
}
\source{
See Cao et. al. 2025 for details. TO-DO-LIST
}
\usage{
Weaker_GWAS_Effect
}
\description{
An example dataset with simulated statistics and LD for 5 traits
}
\seealso{
Other colocboost_data: 
\code{\link{Heterogeneous_Effect}},
\code{\link{Ind_5traits}},
\code{\link{Non_Causal_Strongest_Marginal}},
\code{\link{Sumstat_5traits}}
}
\concept{colocboost_data}
\keyword{datasets}
